refactor: validateDirectionsContinuity 책임 Line -> Directions 생성자로 이동
