+++ src/main/java/nextstep/ladder/domain/Line.java
@@ -1,11 +1,8 @@
 package nextstep.ladder.domain;
 
-import nextstep.ladder.exception.LineDuplicateException;
 import nextstep.ladder.strategy.DirectionStrategy;
-
 import java.util.*;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import static nextstep.ladder.domain.Direction.*;
 
@@ -30,17 +27,7 @@ public class Line {
     }
 
     public Line(List<Direction> directions) {
-        validateDirectionsContinuity(directions);
         this.directions = new Directions(directions);
-    }
-
-    private void validateDirectionsContinuity(List<Direction> directions) {
-        IntStream.range(0, directions.size() - 1)
-                .filter(index -> directions.get(index).isLeft() && directions.get(index + 1).isLeft())
-                .findFirst()
-                .ifPresent(i -> {
-                    throw new LineDuplicateException();
-                });
     }
 
     private static List<Direction> createDirections(DirectionStrategy directionStrategy, int countOfPerson) {
