+++ src/test/java/nextstep/ladder/domain/DirectionsTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.exception.LineDuplicateException;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -12,6 +14,22 @@ import static nextstep.ladder.domain.Direction.*;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 class DirectionsTest {
+
+    @ParameterizedTest
+    @MethodSource("provideDirections")
+    @DisplayName("실패 - 가로 라인이 겹치는 경우 예외가 발생한다.")
+    void fail_line_duplicate(List<Direction> directions) {
+        Assertions.assertThatThrownBy(() -> new Directions(directions))
+                .isInstanceOf(LineDuplicateException.class)
+                .hasMessage("가로 라인이 겹칩니다.");
+    }
+
+    private static Stream<Arguments> provideDirections() {
+        return Stream.of(
+                Arguments.of(List.of(RIGHT, LEFT, LEFT, STAY)),
+                Arguments.of(List.of(STAY, RIGHT, LEFT, LEFT))
+        );
+    }
 
     @ParameterizedTest
     @MethodSource("provideDirection")
